+++ CHANGELOG.md
@@ -33,3 +33,7 @@
 - **FEATURE**: Added new function to request location permission in macos
 - **FEATURE**: Updated example app to request location permission in macos
 
+## 0.0.9
+
+- **BUGFIX**: Fixed request location permission in macos
+- **BUGFIX**: Fixed error in Windows when the SDK installed is not updated and the enum pBssEntry->dot11BssPhyType does not contain all possible values of the switch/case